+++ src/utils/index.ts
@@ -5,3 +5,4 @@ export * from "./color";
 export * from "./rotation";
 export * from "./pressure";
 export * from "./random";
+export * from "./flow";